dialogflow_api/requirements.txt
fastapi
uvicorn
google-cloud-dialogflow
